+++ Mutagen.Bethesda.Fallout4/Records/Major Records/GlobalBool.cs
@@ -61,7 +61,7 @@ partial class GlobalBoolBinaryOverlay
             .ExtractSubrecordMemory(_recordData, _DataLocation.Value, _package.MetaData.Constants).Float() != 0;
     }
 
-    partial void DataCustomParse(OverlayStream stream, long finalPos, int offset)
+    partial void DataCustomParse(OverlayStream stream, int finalPos, int offset)
     {
         _DataLocation = (ushort)(stream.Position - offset);
     }
