+++ Mutagen.Bethesda.Fallout4/Records/Major Records/GlobalShort.cs
@@ -59,7 +59,7 @@ partial class GlobalShortBinaryOverlay
         if (!_DataLocation.HasValue) return default;
         return (short)HeaderTranslation.ExtractSubrecordMemory(_recordData, _DataLocation.Value, _package.MetaData.Constants).Float();
     }
-    partial void DataCustomParse(OverlayStream stream, long finalPos, int offset)
+    partial void DataCustomParse(OverlayStream stream, int finalPos, int offset)
     {
         _DataLocation = (ushort)(stream.Position - offset);
     }
