+++ Mutagen.Bethesda.Fallout4/Records/Major Records/MagicEffect.cs
@@ -183,7 +183,7 @@ partial class MagicEffectBinaryOverlay
 {
     public IReadOnlyList<IConditionGetter> Conditions { get; private set; } = Array.Empty<IConditionGetter>();
 
-    partial void ConditionsCustomParse(OverlayStream stream, long finalPos, int offset, RecordType type, PreviousParse lastParsed)
+    partial void ConditionsCustomParse(OverlayStream stream, int finalPos, int offset, RecordType type, PreviousParse lastParsed)
     {
         Conditions = ConditionBinaryOverlay.ConstructBinayOverlayList(stream, _package);
     }
