+++ Mutagen.Bethesda.Fallout4/Records/Major Records/ScenePhase.cs
@@ -53,12 +53,12 @@ partial class ScenePhaseBinaryOverlay
 
     public IReadOnlyList<IConditionGetter> CompletionConditions { get; private set; } = Array.Empty<IConditionGetter>();
 
-    partial void StartConditionsCustomParse(OverlayStream stream, long finalPos, int offset, RecordType type, PreviousParse lastParsed)
+    partial void StartConditionsCustomParse(OverlayStream stream, int finalPos, int offset, RecordType type, PreviousParse lastParsed)
     {
         StartConditions = ConditionBinaryOverlay.ConstructBinayOverlayList(stream, _package);
     }
 
-    partial void CompletionConditionsCustomParse(OverlayStream stream, long finalPos, int offset, RecordType type, PreviousParse lastParsed)
+    partial void CompletionConditionsCustomParse(OverlayStream stream, int finalPos, int offset, RecordType type, PreviousParse lastParsed)
     {
         stream.ReadSubrecord(RecordTypes.NEXT);
         CompletionConditions = ConditionBinaryOverlay.ConstructBinayOverlayList(stream, _package);
