+++ Mutagen.Bethesda.Oblivion/Records/Major Records/Water.cs
@@ -94,7 +94,7 @@ partial class WaterBinaryOverlay
 {
     private WaterData? _waterData;
 
-    partial void DataCustomParse(OverlayStream stream, long finalPos, int offset)
+    partial void DataCustomParse(OverlayStream stream, int finalPos, int offset)
     {
         this._waterData = WaterBinaryCreateTranslation.CreateCustom(new MutagenFrame(new MutagenInterfaceReadStream(stream, _package.MetaData)));
     }
