+++ Mutagen.Bethesda.Oblivion/Records/Major Records/WorldspaceSubBlock.cs
@@ -9,7 +9,7 @@ internal partial class WorldspaceSubBlockBinaryOverlay
 {
     public IReadOnlyList<ICellGetter> Items { get; private set; } = Array.Empty<ICellGetter>();
 
-    partial void ItemsCustomParse(OverlayStream stream, long finalPos, int offset, RecordType type, PreviousParse lastParsed)
+    partial void ItemsCustomParse(OverlayStream stream, int finalPos, int offset, RecordType type, PreviousParse lastParsed)
     {
         this.Items = BinaryOverlayList.FactoryByArray<ICellGetter>(
             mem: stream.RemainingMemory,
