+++ Mutagen.Bethesda.Skyrim/Records/Major Records/CameraPath.cs
@@ -56,7 +56,7 @@ partial class CameraPathBinaryOverlay
 
     public bool ZoomMustHaveCameraShots => _zoomLoc.HasValue && HeaderTranslation.ExtractSubrecordMemory(_recordData, _zoomLoc.Value, _package.MetaData.Constants)[0] < 128;
 
-    partial void ZoomCustomParse(OverlayStream stream, long finalPos, int offset)
+    partial void ZoomCustomParse(OverlayStream stream, int finalPos, int offset)
     {
         _zoomLoc = stream.Position - offset;
     }
