+++ Mutagen.Bethesda.Starfield/Records/Major Records/ColorRecord.cs
@@ -136,7 +136,7 @@ partial class ColorRecordBinaryOverlay
 
     partial void DataCustomParse(
         OverlayStream stream,
-        long finalPos,
+        int finalPos,
         int offset)
     {
         _cnamLocation = (stream.Position - offset);
@@ -144,7 +144,7 @@ partial class ColorRecordBinaryOverlay
 
     partial void FlagsCustomParse(
         OverlayStream stream,
-        long finalPos,
+        int finalPos,
         int offset)
     {
         _fnamLocation = (stream.Position - offset);
