+++ Mutagen.Bethesda.Starfield/Records/Major Records/FindMatchingRefNearAlias.cs
@@ -41,7 +41,7 @@ partial class FindMatchingRefNearAliasBinaryOverlay
     int? _aliasIDLoc;
     public partial Int16? GetAliasIDCustom() => _aliasIDLoc == null ? default(short?) : checked((short)BinaryPrimitives.ReadInt32LittleEndian(HeaderTranslation.ExtractSubrecordMemory(_recordData, _aliasIDLoc.Value, _package.MetaData.Constants)));
 
-    partial void AliasIDCustomParse(OverlayStream stream, long finalPos, int offset)
+    partial void AliasIDCustomParse(OverlayStream stream, int finalPos, int offset)
     {
         _aliasIDLoc = stream.Position - offset;
     }
